# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/home": {
            "get": {
                "description": "Admin dashbord",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "Admin dashbord",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.AdminDashboard"
                        }
                    }
                }
            }
        },
        "/admin/loginpassword": {
            "post": {
                "description": "Admin login with password and phone number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "Admin Login with password",
                "parameters": [
                    {
                        "description": "Admin Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                }
            }
        },
        "/admin/registeradmin": {
            "post": {
                "description": "Adding new admin to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "registering new admin",
                "parameters": [
                    {
                        "description": "Admin Data",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Admin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Admin"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "description": "Deleting cookie from the browser while logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "logout",
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/service1/healthcheck": {
            "get": {
                "description": "Service Health Check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "admin\u0026user Authentication service health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/service2/healthcheck": {
            "get": {
                "description": "Service Health Check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "admin dashboard service health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/loginwithotp": {
            "post": {
                "description": "Login for user with otp",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone No",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                }
            }
        },
        "/user/loginwithpassword": {
            "post": {
                "description": "Login for user with password",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone No",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                }
            }
        },
        "/user/otpvalidation": {
            "post": {
                "description": "Validating user otp for login validation",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "Otp Validation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Otp",
                        "name": "otp",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key",
                        "name": "key",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "phone",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "resend",
                        "name": "resend",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Login"
                        }
                    }
                }
            }
        },
        "/user/signup": {
            "post": {
                "description": "Adding new user to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "signup",
                "parameters": [
                    {
                        "description": "User Data",
                        "name": "userInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Signup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Signup"
                        }
                    }
                }
            }
        },
        "/user/signupotpvalidation": {
            "post": {
                "description": "Validating user otp for signup",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "Sign Up Otp Validation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Twilio Key",
                        "name": "key",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Otp",
                        "name": "otp",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/signupwithotp": {
            "post": {
                "description": "Adding new user to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User\u0026Admin Authentication"
                ],
                "summary": "signup with opt validation",
                "parameters": [
                    {
                        "description": "User Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Signup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Admin": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "adminname": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "entity.AdminDashboard": {
            "type": "object",
            "properties": {
                "averageordervalue": {
                    "type": "integer"
                },
                "newusers": {
                    "type": "integer"
                },
                "pendingorders": {
                    "type": "integer"
                },
                "returnorders": {
                    "type": "integer"
                },
                "stocklesscategory": {
                    "type": "string"
                },
                "totalorders": {
                    "type": "integer"
                },
                "totalproducts": {
                    "type": "integer"
                },
                "totalquery": {
                    "type": "integer"
                },
                "totalrevenue": {
                    "type": "integer"
                },
                "totalusers": {
                    "type": "integer"
                }
            }
        },
        "entity.Login": {
            "type": "object",
            "required": [
                "password",
                "phone"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "entity.Signup": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname",
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "wallet": {
                    "type": "integer"
                }
            }
        }
    }
}